1.3.0
Uppgiften är nu klar.
Har fixat alla VG mål.

Man kan nu markera ifall en todo är klar eller inte. Stylingen på todon ändras utefter detta.

Limitar också hur många resultat ifrån "databasen" man får till 10.

+++ script.js
@@ -8,11 +8,12 @@ let todosdiv = document.getElementById("todosdiv");
 let todos = [];
 
 
+
+// Limit till 10 resultat
 const fetchTodos =  async () => {
-  const response = await fetch('https://jsonplaceholder.typicode.com/todos');
+  const response = await fetch('https://jsonplaceholder.typicode.com/todos?_limit=10');
   const data = await response.json();
   todos = data;
-  console.log(todos)
   printTodos();
 }
 
@@ -22,7 +23,7 @@ fetchTodos();
 //#region InputValidationAndError
 
 function isInputEmpty(){
-  if(input.value === '' || input.value == null)
+  if(input.value.trim() === '' || input.value == null)
   {return true;}
   return false;
 }
@@ -36,27 +37,55 @@ function printError(){
 }
 //#endregion
 
-
+// Skapar Elementet som innehåller värden av todo objektet ifrån todos.
 function createTodoElement(todo){
 
   let card = document.createElement("div");
   card.classList.add("todo");
-
+  // Skapar paragrpahen ger den värdet och kollar ifall todon är klar eller inte. Lägger till css classen beroende på resultatet
   let paragraph = document.createElement("p");
   paragraph.innerHTML = todo.title;
+  
+  if(todo.completed == true)
+  paragraph.classList.add("completedPara");
+
+  let btnDiv = document.createElement("div");
+  btnDiv.classList.add("btnDiv");
+
+  let statusBtn = document.createElement("button");
+  statusBtn.innerHTML = "O";
+  statusBtn.classList.add("statusBtn");
+
+  if(todo.completed == false)
+  statusBtn.classList.add("unfinishedStatusBtn");
+  else
+  statusBtn.classList.add("completedStatusBtn");
 
   let delBtn = document.createElement("button");
   delBtn.innerHTML = "X";
   delBtn.classList.add("deleteBtn");
 
-  delBtn.addEventListener('click', () => removeTodo(todo.id, card))
+  if(todo.completed == false)
+  delBtn.classList.add("hiddenBtn");
+
+ 
+
+  
+  delBtn.addEventListener('click', () => removeTodo(todo))
+  statusBtn.addEventListener('click', () => completeTodo(todo, card))
+  
 
   card.appendChild(paragraph);
-  card.appendChild(delBtn);
+  card.appendChild(btnDiv);
+
+  btnDiv.appendChild(statusBtn);
+  btnDiv.appendChild(delBtn);
+  
+  
 
   return card;
 }
-
+// Tar bort allt innehåll och skriver sedan ut allt innehåll. Kallar på createTodoElementet och skickar in todon.
 function printTodos(){
   todosdiv.innerHTML = "";
   todos.forEach(todo =>
@@ -72,7 +101,8 @@ const postTodo = (title) => {
       'Content-type': 'application/json; charset=UTF-8'
     },
     body: JSON.stringify({
-      title
+      title: title,
+      completed: false
     })
   })
   .then(res => res.json())
@@ -82,25 +112,59 @@ const postTodo = (title) => {
   })
 }
 
-function removeTodo(id, input) {
-  todos = todos.filter(todo => todo.id !== id)
+function removeTodo(inputTodo) {
+  //Javascript check ifall todon är complete
+  if(inputTodo.completed == true)
+  {
+    todos = todos.filter(todo => todo.id !== inputTodo.id)
   
-  printTodos();
-  fetch(`https://jsonplaceholder.typicode.com/todos/${id}`,
-  { method: "DELETE" })
-  .catch(error => console.error(error));
+    printTodos();
+    fetch(`https://jsonplaceholder.typicode.com/todos/${inputTodo.id}`,
+    { method: "DELETE" })
+    .catch(error => console.error(error));
+
+  }
+  
  
+}
+// Kollar ifall todon är complete eller inte. Sätter värdet till det motsata och kollar sedan ifall todon är complete eller inte.
+//  Sätter sedan olika css classer till olika element i "cardet"
+function completeTodo(todo,card) {
+  
+  todo.completed = !todo.completed;
+  if(todo.completed == true)
+  {
+    // refererar till deletebutton
+    card.childNodes[1].childNodes[1].classList.remove("hiddenBtn");
+    // refererar till statusButton
+    card.childNodes[1].childNodes[0].classList.remove("unfinishedStatusBtn");
+    card.childNodes[1].childNodes[0].classList.add("completedStatusBtn");
+    // refererar till card paragrafen
+    card.childNodes[0].classList.add("completedPara");
+  }
+  else{
+    // refererar till deletebutton
+    card.childNodes[1].childNodes[1].classList.add("hiddenBtn");
+    // refererar till statusButton
+    card.childNodes[1].childNodes[0].classList.add("unfinishedStatusBtn");
+    card.childNodes[1].childNodes[0].classList.remove("completedStatusBtn");
+    // refererar till card paragrafen
+    card.childNodes[0].classList.remove("completedPara");
+   
+
+  }
+
 }
 
 
 // Kallas när submit knappen trycks ner
 form.addEventListener("submit", e =>{
   // Kollar ifall det är tomt och isåfall kallar kallar på printError(), printError ändrar sedan texten 
-  e.preventDefault();
+  
   if(isInputEmpty())
   {
     printError();
-    e.preventDefault();
+    
   }
   else{
     clearErrors();
@@ -108,7 +172,7 @@ form.addEventListener("submit", e =>{
     input.value = "";
   }
   e.preventDefault();
-  console.log("reee");
+  
   
 })
